add release cron task
