+++ .github/workflows/desktop-release.yml
@@ -2,8 +2,10 @@ name: Desktop Release
 
 on:
   push:
+    tags:
+      - 'v*'  # This will trigger the workflow on tags starting with 'v'
     branches:
-      - release
+      - master
 
 concurrency:
   group: ${{ github.workflow }}-${{ github.ref }}
@@ -16,42 +18,42 @@ jobs:
   get-version:
     runs-on: ubuntu-latest
     outputs:
-      version: ${{ steps.get-version.outputs.version }}
+      version: ${{ steps.vars.outputs.version }}
     steps: 
       - uses: actions/checkout@v3
 
-      - name: Run Version Script
-        id: get-version
+      - name: Get Tag
+        id: vars
         run: |
-          echo "version=$(node scripts/get-app-version.js)" >> "$GITHUB_OUTPUT"
-  check-release:
-    needs: [get-version]
-    runs-on: ubuntu-latest
+          echo "version=${GITHUB_REF#refs/tags/}" >> "$GITHUB_OUTPUT"
+  # check-release:
+  #   needs: [get-version]
+  #   runs-on: ubuntu-latest
 
-    steps:
-      - name: Checkout Repository
-        uses: actions/checkout@v2
+  #   steps:
+  #     - name: Checkout Repository
+  #       uses: actions/checkout@v3
 
-      - name: Check Release
-        id: check_release
-        run: |
-          RELEASE_NAME='${{ needs.get-version.outputs.version }}'
+  #     - name: Check Release
+  #       id: check_release
+  #       run: |
+  #         RELEASE_NAME='${{ needs.get-version.outputs.version }}'
 
-          RELEASE_EXISTS=$(curl -s -H "Authorization: Bearer ${{ secrets.GITHUB_TOKEN }}" \
-            "https://api.github.com/repos/${{ github.repository }}/releases" | \
-            jq ".[] | select(.tag_name == \"$RELEASE_NAME\")")
+  #         RELEASE_EXISTS=$(curl -s -H "Authorization: Bearer ${{ secrets.GITHUB_TOKEN }}" \
+  #           "https://api.github.com/repos/${{ github.repository }}/releases" | \
+  #           jq ".[] | select(.tag_name == \"$RELEASE_NAME\")")
 
-          if [ -z "$RELEASE_EXISTS" ]; then
-            echo "Release does not exist."
-          else
-            echo "Release already exists."
-            exit 1
-          fi
+  #         if [ -z "$RELEASE_EXISTS" ]; then
+  #           echo "Release does not exist."
+  #         else
+  #           echo "Release already exists."
+  #           exit 1
+  #         fi
   build-binaries:
     name: Build (${{ matrix.config.os }})
     runs-on: ${{ matrix.config.os }}
     # if: startsWith(github.ref, 'refs/tags/')
-    needs: [get-version, check-release]
+    needs: [get-version]
     strategy:
       matrix:
         config:
@@ -64,9 +66,9 @@ jobs:
           - os: ubuntu-latest
             goarch: amd64
             target: x86_64-unknown-linux-gnu
-          # - os: windows-latest
-          #   goarch: amd64
-          #   target: x86_64-pc-windows-msvc
+          - os: windows-latest
+            goarch: amd64
+            target: x86_64-pc-windows-msvc
     steps:
       - name: Checkout
         uses: actions/checkout@v1
@@ -115,9 +117,10 @@ jobs:
       #   if: matrix.config.os == 'ubuntu-latest' && startsWith(github.ref, 'refs/tags/')
       #   run: sudo apt install flatpak flatpak-builder elfutils 
           
-      - name: Build App Package
+      - name: Build App Package (Unix)
+        if: matrix.config.os != 'windows-latest'
         run: |
-          TAMAGUI_TARGET=web yarn electron:make
+          TAMAGUI_TARGET=web yarn desktop:make
         env:
           APP_VERSION: '${{ needs.get-version.outputs.version }}'
           GITHUB_TOKEN: '${{ secrets.GITHUB_TOKEN }}'
@@ -125,6 +128,16 @@ jobs:
           APPLE_ID: ${{ secrets.APPLE_ID }}
           APPLE_ID_PASSWORD: '${{ secrets.APPLE_ID_PASSWORD }}'
           APPLE_TEAM_ID: '${{ secrets.APPLE_TEAM_ID }}'
+
+      - name: Build App Package (Win32)
+        if: matrix.config.os == 'windows-latest'
+        run: |
+          $env:TAMAGUI_TARGET="web"
+          yarn desktop:make
+        env:
+          APP_VERSION: '${{ needs.get-version.outputs.version }}'
+          GITHUB_TOKEN: '${{ secrets.GITHUB_TOKEN }}'
+          SENTRY_AUTH_TOKEN: ${{ secrets.SENTRY_AUTH_TOKEN }}
           
           # WINDOWS_PFX_FILE: '${{ steps.write_file.outputs.filePath }}'
           # WINDOWS_PFX_PASSWORD: '${{ secrets.WINDOWS_PFX_PASSWORD }}'
