+++ .github/workflows/release-tag.yml
@@ -1,0 +1,36 @@
+name: Weekly Release Tag
+
+on:
+  schedule:
+    - cron: '0 6 * * 2'  # Runs every Tuesday at 6am
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: Checkout repository
+        uses: actions/checkout@v2
+
+      - name: Install Node.js 20
+        uses: actions/setup-node@v3
+        with:
+          node-version: 20
+
+      - name: Get new version
+        run: |
+          node scripts/get-app-version.mjs > version.txt
+
+      - name: Read version from file
+        id: version
+        run: echo "version=$(cat version.txt)" >> "$GITHUB_OUTPUT"
+
+      - name: Create and push tag
+        env:
+          GITHUB_TOKEN: '${{ secrets.GITHUB_TOKEN }}'
+        run: |
+          git config --global user.email "actions@github.com"
+          git config --global user.name "GitHub Actions"
+          git tag ${{ steps.version.outputs.version }}
+          git push origin ${{ steps.version.outputs.version }}
+
